[Change]配餐逻辑修改

+++ 餐型匹配/餐型匹配/数据/Meal.h
@@ -26,7 +26,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic,strong) Food           *food;
 
 //Meal ID
-@property (nonatomic,assign) int         ID;
+@property (nonatomic,assign) NSUInteger     ID;
 @end
 
 NS_ASSUME_NONNULL_END
